20191127-1722解析版本文件方法
1.解析版本文件方法
2.更改ab包size类型

+++ Assets/YouYouFramework/Managers/Resource/AssetBundleInfoEntity.cs
@@ -18,7 +18,7 @@ public class AssetBundleInfoEntity
     /// <summary>
     /// AB包大小
     /// </summary>
-    public int Size;
+    public ulong Size;
     /// <summary>
     /// 是否为初始化资源
     /// </summary>
